+++ NINA.Equipment/Equipment/AscomDevice.cs
@@ -310,8 +310,12 @@ namespace NINA.Equipment.Equipment {
         /// <typeparam name="PropT"></typeparam>
         /// <param name="propertyName">Property Name of the AscomDevice property</param>
         /// <param name="defaultValue">The default value to be returned when not connected or not implemented</param>
+        /// <param name="cacheInterval">The minimum interval between actual polls to the device</param>
+        /// <param name="rethrow">When set - any error will be rethrown and not handled internally</param>
+        /// <param name="useLastKnownValueOnError">When rethrow is false and this is set, the last known value will be used as a fallback. Otherwise the errorValue parameter will be used</param>
+        /// <param name="errorValue">The value to be returned when rethrow and useLastKnownValueOnError are both set to false and an error occurs during reading of the property</param>
         /// <returns></returns>
-        protected PropT GetProperty<PropT>(string propertyName, PropT defaultValue, TimeSpan? cacheInterval = null, bool rethrow = false) {            
+        protected PropT GetProperty<PropT>(string propertyName, PropT defaultValue, TimeSpan? cacheInterval = null, bool rethrow = false, bool useLastKnownValueOnError = true, PropT errorValue = default) {            
             if (device != null) {
                 var type = device.GetType();
 
@@ -378,11 +382,12 @@ namespace NINA.Equipment.Equipment {
                     }
                 }
 
-                var val = (PropT)memory.LastValue;
+                // Polling the property failed for all retries
+                var val = useLastKnownValueOnError ? (PropT)memory.LastValue : errorValue;
                 if (memory.ConsecutiveErrors > memory.ConsecutiveErrorThreshold) {
-                    Logger.Trace($"GET {type.Name}.{propertyName} failed - Returning last known value {val}");
+                    Logger.Trace($"GET {type.Name}.{propertyName} failed - Returning {(useLastKnownValueOnError ? "last known" : "error")} value {val}");
                 } else {
-                    Logger.Info($"GET {type.Name}.{propertyName} failed - Returning last known value {val}");
+                    Logger.Info($"GET {type.Name}.{propertyName} failed - Returning {(useLastKnownValueOnError ? "last known" : "error")} value {val}");
                 }   
                 return val;
             }