+++ RELEASE_NOTES.md
@@ -16,6 +16,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
   - However, please note that you must close the setup screen for the specific device before being able to connect to it.
 - Fixed that an incorrect flag was evaluated for the mount capability of PulseGuiding
 - Fixed an issue in TimeCondition where the time would be overridden to current time
+- When polling the SafetyMontior for "IsSafe" and an error occurs it will now always assume unsafe conditions instead of using the last known value
 
 # Version 3.0